pyproject.toml
[project]
name = "ai-assignment-grader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "fastmcp>=2.14.5",
    "fuzzywuzzy>=0.18.0",
    "google-generativeai>=0.8.6",
    "openai>=1.78.0",
    "pymupdf>=1.25.5",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "python-levenshtein>=0.27.1",
    "requests>=2.32.3",
    "streamlit>=1.45.0",
    "uvicorn>=0.41.0",
]
